requirements.txt
streamlit
google-generativeai
python-dotenv
pandas
openpyxl
PyPDF2
python-docx